analog: separated analog components into separate file
