+++ BinPy/analog/__init__.py
@@ -1,2 +1,4 @@
 from BinPy.analog.source import *
 from BinPy.analog.base import *
+from BinPy.analog.components import *
+from BinPy.analog.converters import *
